autoland-web-api: hack around docker-compose mounting over autolandweb.egg-info. r=mars
When installing a python package in development mode a <package>.egg-info
directory is created in the packages folder. Since docker-compose mounts
over the code directory when starting the image the egg-info folder is
hidden.

To work around this we create the autolandweb.egg-info directory in the
root of the image and add the app directory to the PYTHONPATH so it can
be found. This is pretty hacky but should only be required for the
development container, not production images.