autoland-ui: add UI for landing to Try. (bug 1334225). r=davidwalsh,smacleod
Add the UI elements and styles to support landing to try.
In the future we will need to update the appropriate callbacks
to actually hit the API and land to Try. Also we will need to
add a small widget to show the status of the try attempt.

For now the UI slides out from the right, allowing partial view
of the commit log. The UI allows for manually entering a try string
or using the graphical chooser embedded from
https://mozilla-releng.net/trychooser/?embed. Our CSP policies
must allow this domain.

MozReview-Commit-ID: P8NYw3Mf0q

--HG--
extra : rebase_source : dfd87661c69b0c16b91d6ce97794cf81717b54ca