autoland: introduce a circle.yml for building production on CircleCI.
We adapt a circle.yml from some of the other dockerflow projects,
but since we are going to be building multiple images on each push
we need to customize things a bit. In the future most of these
build steps should be factored out into a generic script we can
use to build each service in a single call.