autoland-webapi: add universal mozlog format logging.
This change introduces a new MozLogFormatter(logging.formatter)
class and configuration of the root logger to use it. Now any
call to a pythong logging function will properly output mozlog.

A PrettyMozLogFormatter is also introduced which will serialize
the mozlog json in a multi-line pretty format. This formatter
is used in the dev environment and when autolandweb is passed
the --pretty-log or the AUTOLANDWEB_PRETTY_LOG environment
variable is set.

--HG--
extra : amend_source : 104d486438ce33f997ea2a0d2093070730b00e38