autoland-ui: separate ui to access api from its own host (bug 1341196). r=davidwalsh,mars,smacleod
Now that the API is on it's own host, this updates the UI to use access
it via that host. When in the development environment it will use the
same hostname that the UI was served from and the port 9999 to access the
API. When in production, it will use the actual host of the production
API server.

MozReview-Commit-ID: KXEVH07MeVo

--HG--
extra : rebase_source : f44b064dd998c6f069432498989c9ea2cf6a09eb
extra : amend_source : 032d9c894beb83a0781692a8a1566b2fc34088fe