autoland: add nginx config and port to see how ui is being served from compiled files instead of the dev server (bug 1344004). r=smacleod
During normal development it is better to proxy to the UI dev server to
work since the ui dev server will automatically rebuild files, give
better error messages, and allow use of the react dev tools in chrome.
When the devserver runs it builds the files exactly the same as 'yarn build'
but it also serves some extra files. 99% of the time this isn't a problem.

This commit allows us to double check how the ui works when it is served
directly from nginx from the pre-compiled files - exactly as it happens
in production. This is especially useful in verifying that files that
should *not* be compiled are copied and served correctly.

MozReview-Commit-ID: 3RbnrEKdGZa

--HG--
extra : rebase_source : 36cd466db17dacf8c709c6e6d31e89a5456991fa