autoland-webapi: remove assumption of multiple CORS origins.
We were mistakenly assuming you could specify multiple origins in the
Access-Control-Allow-Origin header, which should only contain a single
origin or "*". The proper way to allow multiple origins is to inspect
the incoming Origin header, and compare it against a list of acceptable
origins, Echoing the origin back in the Access-Control-Allow-Origin
header if there is a match.

While we might require multiple origins in the future we don't require
it currently, so properly handling more than one will be left to another
change.

--HG--
extra : amend_source : 65f0c4266153fde8a465b876929ee3c64698c3e3